+++ geemap/eefolium.py
@@ -515,7 +515,7 @@ class Map(folium.Map):
         attribution=".",
         opacity=1.0,
         shown=True,
-        titiler_endpoint="https://api.cogeo.xyz/",
+        titiler_endpoint="https://titiler.xyz",
         **kwargs,
     ):
         """Adds a COG TileLayer to the map.
@@ -526,10 +526,10 @@ class Map(folium.Map):
             attribution (str, optional): The attribution to use. Defaults to '.'.
             opacity (float, optional): The opacity of the layer. Defaults to 1.
             shown (bool, optional): A flag indicating whether the layer should be on by default. Defaults to True.
-            titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://api.cogeo.xyz/".
+            titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://titiler.xyz".
         """
-        tile_url = get_cog_tile(url, titiler_endpoint, **kwargs)
-        center = get_cog_center(url, titiler_endpoint)  # (lon, lat)
+        tile_url = cog_tile(url, titiler_endpoint, **kwargs)
+        center = cog_center(url, titiler_endpoint)  # (lon, lat)
         self.add_tile_layer(
             tiles=tile_url,
             name=name,
@@ -546,7 +546,7 @@ class Map(folium.Map):
         attribution=".",
         opacity=1.0,
         shown=True,
-        titiler_endpoint="https://api.cogeo.xyz/",
+        titiler_endpoint="https://titiler.xyz",
         username="anonymous",
         overwrite=False,
         show_footprints=False,
@@ -561,14 +561,14 @@ class Map(folium.Map):
             attribution (str, optional): The attribution to use. Defaults to '.'.
             opacity (float, optional): The opacity of the layer. Defaults to 1.
             shown (bool, optional): A flag indicating whether the layer should be on by default. Defaults to True.
-            titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://api.cogeo.xyz/".
+            titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://titiler.xyz".
             username (str, optional): The username to create mosaic using the titiler endpoint. Defaults to 'anonymous'.
             overwrite (bool, optional): Whether or not to replace existing layer with the same layer name. Defaults to False.
             show_footprints (bool, optional): Whether or not to show footprints of COGs. Defaults to False.
             verbose (bool, optional): Whether or not to print descriptions. Defaults to True.
         """
         layername = name.replace(" ", "_")
-        tile = get_cog_mosaic(
+        tile = cog_mosaic(
             links,
             titiler_endpoint=titiler_endpoint,
             username=username,
@@ -585,7 +585,7 @@ class Map(folium.Map):
                 )
             coords = []
             for link in links:
-                coord = get_cog_bounds(link)
+                coord = cog_bounds(link)
                 if coord is not None:
                     coords.append(coord)
             fc = coords_to_geojson(coords)
@@ -602,7 +602,7 @@ class Map(folium.Map):
             if verbose:
                 print("The footprint layer has been added.")
         else:
-            center = get_cog_center(links[0], titiler_endpoint)
+            center = cog_center(links[0], titiler_endpoint)
 
         self.set_center(center[0], center[1], zoom=6)
 
@@ -614,7 +614,7 @@ class Map(folium.Map):
         attribution=".",
         opacity=1.0,
         shown=True,
-        titiler_endpoint="https://api.cogeo.xyz/",
+        titiler_endpoint="https://titiler.xyz",
         **kwargs,
     ):
         """Adds a STAC TileLayer to the map.
@@ -625,10 +625,10 @@ class Map(folium.Map):
             attribution (str, optional): The attribution to use. Defaults to '.'.
             opacity (float, optional): The opacity of the layer. Defaults to 1.
             shown (bool, optional): A flag indicating whether the layer should be on by default. Defaults to True.
-            titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://api.cogeo.xyz/".
+            titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://titiler.xyz".
         """
-        tile_url = get_stac_tile(url, bands, titiler_endpoint, **kwargs)
-        center = get_stac_center(url, titiler_endpoint)
+        tile_url = stac_tile(url, bands, titiler_endpoint, **kwargs)
+        center = stac_center(url, titiler_endpoint)
         self.add_tile_layer(
             tiles=tile_url,
             name=name,